Editorial: Move this spec to ARIA monorepo (#17)
In order to more easily keep track of the many ARIA specifications and spec changes, we are moving to an ARIA monorepo. We will merge this PR after [this PR to add the spec to ARIA](https://github.com/w3c/aria/pull/2190).

This repo will remain open for issue tracking, and the editor's draft will still be published here to maintain editor's draft URLs.